refactor: move worker agent tool definitions into dedicated file
